installer: don't reset shell when uninstall
If there's no `~/.shell.pre-oh-my-zsh`, don't assume the default choice
is Bash.  In fact Zsh is the default shell for macOS since Catalina
(10.15) [1], yet users of other OSes have likely to have Bash as their
default.

This commit fix issue #8252

[1] https://support.apple.com/en-us/HT208050

+++ tools/uninstall.sh
@@ -25,12 +25,8 @@ if [ -e "$ZSHRC_ORIG" ]; then
   echo "Your original zsh config was restored."
 fi
 
-if hash chsh >/dev/null 2>&1; then
-  if [ -f ~/.shell.pre-oh-my-zsh ]; then
-    old_shell=$(cat ~/.shell.pre-oh-my-zsh)
-  else
-	old_shell=/bin/bash
-  fi
+if hash chsh >/dev/null 2>&1 && [ -f ~/.shell.pre-oh-my-zsh ]; then
+  old_shell=$(cat ~/.shell.pre-oh-my-zsh)
   echo "Switching your shell back to '$old_shell':"
   if chsh -s "$old_shell"; then
     rm -f ~/.shell.pre-oh-my-zsh
